refactor: 전반적인 디렉토리 구조 수정
